requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
boto=2.49.0=py39h06a4308_0
boto3=1.16.0=py_0
botocore=1.19.1=py_0
bottleneck=1.3.4=py39hce1f21e_0
brotlipy=0.7.0=py39h27cfd23_1003
bz2file=0.98=py39h06a4308_1
ca-certificates=2022.3.29=h06a4308_1
certifi=2021.10.8=py39h06a4308_2
cffi=1.15.0=py39hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.2=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=36.0.0=py39h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cython=0.29.28=py39h295c915_0
docutils=0.15.2=py39h06a4308_1
emoji=1.7.0=pypi_0
filelock=3.6.0=pypi_0
gensim=4.1.2=py39h295c915_0
huggingface-hub=0.5.1=pypi_0
idna=2.10=py_0
importlib-metadata=4.11.3=py39hf3d152e_1
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.0.3=pyhd3eb1b0_0
jmespath=0.10.0=py_0
joblib=0.17.0=py_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libstdcxx-ng=9.3.0=hd4cf53a_17
libuv=1.40.0=h7b6447c_0
markupsafe=2.0.1=py39h27cfd23_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
ncurses=6.3=h7f8727e_2
networkx=2.7.1=pyhd3eb1b0_0
nltk=3.7=pypi_0
numexpr=2.8.1=py39h6abb31d_0
numpy=1.21.5=py39he7a7128_1
numpy-base=1.21.5=py39hf524024_1
openssl=1.1.1n=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=py39h295c915_1
pillow=9.1.0=pypi_0
pip=21.2.4=py39h06a4308_0
prettytable=3.2.0=pyhd8ed1ab_0
pycparser=2.20=py_2
pyg=2.0.4=py39_torch_1.11.0_cu113
pyopenssl=19.1.0=py_1
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.12=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-louvain=0.15=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.11.0=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-cluster=1.6.0=py39_torch_1.11.0_cu113
pytorch-mutex=1.0=cuda
pytorch-scatter=2.0.9=py39_torch_1.11.0_cu113
pytorch-sparse=0.6.13=py39_torch_1.11.0_cu113
pytorch-spline-conv=1.2.1=py39_torch_1.11.0_cu113
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=py39h7f8727e_1
readline=8.1.2=h7f8727e_1
regex=2022.3.15=pypi_0
requests=2.27.1=pyhd3eb1b0_0
s3transfer=0.3.6=pyhd3eb1b0_0
sacremoses=0.0.49=pypi_0
scikit-learn=1.0.2=py39h51133e4_1
scipy=1.7.3=py39hc147768_0
sentence-transformers=2.2.0=pypi_0
sentencepiece=0.1.96=pypi_0
setuptools=61.2.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smart_open=1.9.0=py_0
sqlite=3.38.2=hc218d9a_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.12.1=pypi_0
torchvision=0.12.0=pypi_0
tqdm=4.63.0=pyhd3eb1b0_0
transformers=4.18.0=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.25.11=py_0
wcwidth=0.2.5=pyh9f0ad1d_2
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yacs=0.1.6=pyhd3eb1b0_1
yaml=0.2.5=h7b6447c_0
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.12=h7f8727e_1
